+++ backend/src/payments/dto.ts
@@ -1,7 +1,8 @@
 import { PaymentMethod } from '@prisma/client';
 import { IsDateString, IsEnum, IsOptional, IsString } from 'class-validator';
+import { PaginationQueryDto } from '../common/dto/pagination-query.dto';
 
-export class QueryPaymentDto {
+export class QueryPaymentDto extends PaginationQueryDto {
   @IsOptional()
   @IsDateString()
   date?: string;